# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NamThu Education API",
        "description": "Comprehensive API documentation for NamThu Education Laravel Backend",
        "contact": {
            "email": "support@namthuedu.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/student/tests": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Get student tests",
                "description": "Get list of tests assigned to authenticated student",
                "operationId": "da165138fff6bf43457e22aa4dab5723",
                "responses": {
                    "200": {
                        "description": "Tests retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/student/tests/{id}": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Get test details",
                "description": "Get detailed information about a specific test (without correct answers)",
                "operationId": "a8ac21b5323c89739fcf5dca4d145375",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Test details retrieved successfully"
                    },
                    "404": {
                        "description": "Test not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/student/tests/{id}/start": {
            "post": {
                "tags": [
                    "Students"
                ],
                "summary": "Start test",
                "description": "Start taking a test (creates submission)",
                "operationId": "8fcf332569616778f7ee4eeefa8a5222",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Test started successfully"
                    },
                    "400": {
                        "description": "Cannot start test (already completed, expired, etc.)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/student/tests/{submissionId}/answer": {
            "post": {
                "tags": [
                    "Students"
                ],
                "summary": "Submit answer",
                "description": "Submit answer for a question during test",
                "operationId": "d8528ff6df4b6213d1a22a4a92b46543",
                "parameters": [
                    {
                        "name": "submissionId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "question_id"
                                ],
                                "properties": {
                                    "question_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "answer_id": {
                                        "description": "For multiple choice questions",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "answer_text": {
                                        "description": "For essay questions",
                                        "type": "string",
                                        "example": "Essay answer text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Answer submitted successfully"
                    },
                    "400": {
                        "description": "Invalid submission or question"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/student/tests/{submissionId}/submit": {
            "post": {
                "tags": [
                    "Students"
                ],
                "summary": "Submit test",
                "description": "Submit completed test for grading",
                "operationId": "d45a78274ee9e592aced67ce188e778b",
                "parameters": [
                    {
                        "name": "submissionId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Test submitted successfully"
                    },
                    "400": {
                        "description": "Cannot submit test (not started, already submitted, etc.)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/student/submissions": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Get student submissions",
                "description": "Get submission history for authenticated student",
                "operationId": "c09752d45338f93d3b9c3d545f000fe9",
                "responses": {
                    "200": {
                        "description": "Submissions retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/student/submissions/{id}": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Get submission details",
                "description": "Get detailed information about a specific submission with scores and feedback",
                "operationId": "ca6600fd69c6e514456168f461ad3e70",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Submission details retrieved successfully"
                    },
                    "404": {
                        "description": "Submission not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}